Move embed setup modal state types into the redux store
redux/store/modal.ts imported SdkIframeEmbedSetupModalInitialState and
LegacyStaticEmbeddingModalProps from metabase/plugins, an upward edge.
The types (plus SdkIframeEmbedSetupExperience, which the legacy props
reference) now live with the ModalState they describe, and the plugins
and embedding modules import them from metabase/redux/store/modal.
