+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/components/tests/test-setup.tsx
@@ -15,8 +15,8 @@ import {
 } from "__support__/server-mocks";
 import { mockSettings } from "__support__/settings";
 import { renderWithProviders, waitFor } from "__support__/ui";
-import type { SdkIframeEmbedSetupModalInitialState } from "metabase/plugins";
 import { createMockState } from "metabase/redux/store/mocks";
+import type { SdkIframeEmbedSetupModalInitialState } from "metabase/redux/store/modal";
 import type { Dashboard } from "metabase-types/api";
 import {
   createMockCollection,
