+++ frontend/src/metabase/new/components/NewModals/NewModals.tsx
@@ -9,14 +9,14 @@ import { UpgradeModal } from "metabase/common/components/upsells/components/Upgr
 import { STATIC_LEGACY_EMBEDDING_TYPE } from "metabase/embedding/constants";
 import { PaletteShortcutsModal } from "metabase/palette/components/PaletteShortcutsModal/PaletteShortcutsModal";
 import { useRegisterShortcut } from "metabase/palette/hooks/useRegisterShortcut";
-import type {
-  LegacyStaticEmbeddingModalProps,
-  SdkIframeEmbedSetupModalProps,
-} from "metabase/plugins";
+import type { SdkIframeEmbedSetupModalProps } from "metabase/plugins";
 import { ActionCreator } from "metabase/querying/action-creator";
 import { useDispatch, useSelector } from "metabase/redux";
 import type { State } from "metabase/redux/store";
-import type { ModalState } from "metabase/redux/store/modal";
+import type {
+  LegacyStaticEmbeddingModalProps,
+  ModalState,
+} from "metabase/redux/store/modal";
 import { closeModal, setOpenModal } from "metabase/redux/ui";
 import { useLocation, useNavigate, useParams } from "metabase/router";
 import { Modal, PREVENT_AUTOCOMPLETE_CLIPPING_MODAL_PROPS } from "metabase/ui";
