+++ frontend/src/metabase/plugins/index.ts
@@ -76,9 +76,6 @@ export { PLUGIN_EMBEDDING_IFRAME_SDK } from "./oss/embedding-iframe-sdk";
 export {
   PLUGIN_EMBEDDING_IFRAME_SDK_SETUP,
   type SdkIframeEmbedSetupModalProps,
-  type SdkIframeEmbedSetupModalInitialState,
-  type SdkIframeEmbedSetupExperience,
-  type LegacyStaticEmbeddingModalProps,
 } from "./oss/embedding-iframe-sdk-setup";
 export { PLUGIN_EMBEDDING_SDK } from "./oss/embedding-sdk";
 export {
