+++ frontend/src/metabase/plugins/oss/embedding-iframe-sdk-setup.ts
@@ -1,30 +1,9 @@
-import type { DashboardId } from "metabase-types/api";
+import type { SdkIframeEmbedSetupModalInitialState } from "metabase/redux/store/modal";
 
 export type SdkIframeEmbedSetupModalProps = {
   opened: boolean;
   onClose: () => void;
   initialState?: SdkIframeEmbedSetupModalInitialState;
-};
-
-export type SdkIframeEmbedSetupModalInitialState = {
-  resourceType?: string | null;
-  resourceId?: string | number | null;
-  isGuest?: boolean;
-  useExistingUserSession?: boolean;
-};
-
-export type SdkIframeEmbedSetupExperience =
-  | "dashboard"
-  | "chart"
-  | "exploration"
-  | "browser"
-  | "metabot";
-
-export type LegacyStaticEmbeddingModalProps = {
-  experience: SdkIframeEmbedSetupExperience;
-  dashboardId?: DashboardId | null;
-  questionId?: string | number | null;
-  parentInitialState: SdkIframeEmbedSetupModalInitialState;
 };
 
 const getDefaultPluginEmbeddingIframeSdkSetup = () => ({
